requirements.txt
django-smuggler==0.7.0
py-moneyed==0.6.0
django-money==0.9.1
